# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S.A.L.I.M. Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /objectives:
    get:
      operationId: listObjectives
      tags: [objectives]
      summary: List all objectives for current user
      responses:
        "200":
          description: List of objectives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Objective"
    post:
      operationId: createObjective
      tags: [objectives]
      summary: Create a new objective
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectiveInput"
      responses:
        "201":
          description: Created objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"
  /objectives/active:
    get:
      operationId: getActiveObjective
      tags: [objectives]
      summary: Get the current active objective
      responses:
        "200":
          description: Active objective with full S.A.L.I.M. progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectiveDetail"
        "404":
          description: No active objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /objectives/{id}:
    get:
      operationId: getObjective
      tags: [objectives]
      summary: Get objective by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectiveDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateObjective
      tags: [objectives]
      summary: Update an objective
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectiveUpdate"
      responses:
        "200":
          description: Updated objective
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"
  /objectives/{id}/activate:
    post:
      operationId: activateObjective
      tags: [objectives]
      summary: Set objective as active (archives others)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"

  # ─── S – SPECIFY ────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List weekly reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyReview"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create or update a weekly review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReview"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a weekly review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReview"

  # ─── M – MAINTAIN (HABITS) ──────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── AUTH ──────────────────────────────────────────────
...
    Objective:
      type: object
      required: [id, userId, title, status, currentStep, completedSteps, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, archived]
        currentStep:
          type: string
          enum: [specify, align, layout, implement, maintain]
        completedSteps:
          type: array
          items:
            type: string
        startDate:
          type: ["string", "null"]
          format: date
        targetDate:
          type: ["string", "null"]
          format: date
        progressPercent:
          type: integer
        createdAt:
          type: string
          format: date-time
    ObjectiveDetail:
      type: object
      required: [id, userId, title, status, currentStep, completedSteps, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, archived]
        currentStep:
          type: string
          enum: [specify, align, layout, implement, maintain]
        completedSteps:
          type: array
          items:
            type: string
        startDate:
          type: ["string", "null"]
          format: date
        targetDate:
          type: ["string", "null"]
          format: date
        progressPercent:
          type: integer
        specify:
          $ref: "#/components/schemas/SpecifyData"
        align:
          $ref: "#/components/schemas/AlignData"
        plan:
          $ref: "#/components/schemas/Plan"
        createdAt:
          type: string
          format: date-time
    ObjectiveInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
    ObjectiveUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, completed, archived]
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date

    # ─── SPECIFY ───────────────────────────────────────────
    SpecifyData:
      type: object
      required: [id, objectiveId]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        smartGoal:
          type: ["string", "null"]
        specificWhat:
          type: ["string", "null"]
        measurableHow:
          type: ["string", "null"]
        achievableSteps:
          type: ["string", "null"]
        relevantWhy:
          type: ["string", "null"]
        timeBoundWhen:
          type: ["string", "null"]
        precisionScore:
          type: ["integer", "null"]
        isValidated:
          type: boolean
...
    AlignData:
      type: object
      required: [id, objectiveId]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        deepWhy:
          type: ["string", "null"]
        values:
          type: array
          items:
            type: string
        visualizationNotes:
          type: ["string", "null"]
        alignmentScore:
          type: ["integer", "null"]
        isCompleted:
          type: boolean
...
    Plan:
      type: object
      required: [id, objectiveId, milestones]
      properties:
        id:
          type: integer
        objectiveId:
          type: integer
        summary:
          type: ["string", "null"]
        priorityFocus:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
...
    Milestone:
      type: object
      required: [id, planId, title, weekNumber, isCompleted]
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        weekNumber:
          type: integer
        targetDate:
          type: ["string", "null"]
          format: date
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
...
    WeeklyReview:
      type: object
      required: [id, userId, weekNumber, weekStartDate, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        objectiveId:
          type: ["integer", "null"]
        weekNumber:
          type: integer
        weekStartDate:
          type: string
          format: date
        wins:
          type: ["string", "null"]
        challenges:
          type: ["string", "null"]
        learnings:
          type: ["string", "null"]
        nextWeekFocus:
          type: ["string", "null"]
        energyScore:
          type: ["integer", "null"]
        progressScore:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [weekNumber, weekStartDate]
      properties:
        objectiveId:
          type: integer
        weekNumber:
          type: integer
        weekStartDate:
          type: string
          format: date
        wins:
          type: string
        challenges:
          type: string
        learnings:
          type: string
        nextWeekFocus:
          type: string
        energyScore:
          type: integer
        progressScore:
          type: integer
    ReviewUpdate:
      type: object
      properties:
        wins:
          type: string
        challenges:
          type: string
        learnings:
          type: string
        nextWeekFocus:
          type: string
        energyScore:
          type: integer
        progressScore:
          type: integer

    # ─── HABITS ────────────────────────────────────────────
...