# lib/api-zod/src/generated/types/alignData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S.A.L.I.M. Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AlignData {
  id: number;
  objectiveId: number;
  /** @nullable */
  deepWhy?: string | null;
  values?: string[];
  /** @nullable */
  visualizationNotes?: string | null;
  /** @nullable */
  alignmentScore?: number | null;
  isCompleted?: boolean;
}