# lib/api-zod/src/generated/types/objective.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S.A.L.I.M. Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ObjectiveCurrentStep } from "./objectiveCurrentStep";
import type { ObjectiveStatus } from "./objectiveStatus";

export interface Objective {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ObjectiveStatus;
  currentStep: ObjectiveCurrentStep;
  completedSteps: string[];
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  targetDate?: Date | null;
  progressPercent?: number;
  createdAt: Date;
}