# lib/api-zod/src/generated/types/objectiveDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S.A.L.I.M. Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AlignData } from "./alignData";
import type { ObjectiveDetailCurrentStep } from "./objectiveDetailCurrentStep";
import type { ObjectiveDetailStatus } from "./objectiveDetailStatus";
import type { Plan } from "./plan";
import type { SpecifyData } from "./specifyData";

export interface ObjectiveDetail {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ObjectiveDetailStatus;
  currentStep: ObjectiveDetailCurrentStep;
  completedSteps: string[];
  /** @nullable */
  startDate?: Date | null;
  /** @nullable */
  targetDate?: Date | null;
  progressPercent?: number;
  specify?: SpecifyData;
  align?: AlignData;
  plan?: Plan;
  createdAt: Date;
}