# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * S.A.L.I.M. Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Milestone } from "./milestone";

export interface Plan {
  id: number;
  objectiveId: number;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  priorityFocus?: string | null;
  milestones: Milestone[];
}